impl macro for panel construction
