+++ tools/twix/src/main.rs
@@ -43,6 +43,8 @@ mod repository_parameters;
 mod twix_painter;
 mod value_buffer;
 pub mod visuals;
+pub mod selectable_panel_macro;
+
 
 fn setup_logger() -> Result<(), InitError> {
     Dispatch::new()
@@ -70,109 +72,9 @@ fn main() -> Result<(), eframe::Error> {
     )
 }
 
-#[allow(clippy::large_enum_variant)]
-enum SelectablePanel {
-    BehaviorSimulator(BehaviorSimulatorPanel),
-    Text(TextPanel),
-    Plot(PlotPanel),
-    Image(ImagePanel),
-    ImageSegments(ImageSegmentsPanel),
-    Map(MapPanel),
-    Parameter(ParameterPanel),
-    ManualCalibration(ManualCalibrationPanel),
-    LookAt(LookAtPanel),
-    Remote(RemotePanel),
-    VisionTuner(VisionTunerPanel),
-}
 
-impl SelectablePanel {
-    fn new(nao: Arc<Nao>, value: Option<&Value>) -> Result<SelectablePanel> {
-        let name = value
-            .ok_or(eyre!("Got none value"))?
-            .get("_panel_type")
-            .ok_or(eyre!("value has no _panel_type: {value:?}"))?
-            .as_str()
-            .ok_or(eyre!("_panel_type is not a string"))?;
-        Self::try_from_name(name, nao, value)
-    }
+impl_selectablepanel!(BehaviorSimulatorPanel, TextPanel, PlotPanel, ImagePanel, ImageSegmentsPanel, MapPanel, ParameterPanel, ManualCalibrationPanel, LookAtPanel, RemotePanel, VisionTunerPanel);
 
-    fn try_from_name(name: &str, nao: Arc<Nao>, value: Option<&Value>) -> Result<SelectablePanel> {
-        Ok(match name.to_lowercase().as_str() {
-            "behavior simulator" => {
-                SelectablePanel::BehaviorSimulator(BehaviorSimulatorPanel::new(nao, value))
-            }
-            "text" => SelectablePanel::Text(TextPanel::new(nao, value)),
-            "plot" => SelectablePanel::Plot(PlotPanel::new(nao, value)),
-            "image" => SelectablePanel::Image(ImagePanel::new(nao, value)),
-            "image segments" => SelectablePanel::ImageSegments(ImageSegmentsPanel::new(nao, value)),
-            "map" => SelectablePanel::Map(MapPanel::new(nao, value)),
-            "parameter" => SelectablePanel::Parameter(ParameterPanel::new(nao, value)),
-            "manual calibration" => {
-                SelectablePanel::ManualCalibration(ManualCalibrationPanel::new(nao, value))
-            }
-            "look at" => SelectablePanel::LookAt(LookAtPanel::new(nao, value)),
-            "remote" => SelectablePanel::Remote(RemotePanel::new(nao, value)),
-            "vision tuner" => SelectablePanel::VisionTuner(VisionTunerPanel::new(nao, value)),
-            name => bail!("unexpected panel name: {name}"),
-        })
-    }
-
-    fn save(&self) -> Value {
-        let mut value = match self {
-            SelectablePanel::BehaviorSimulator(panel) => panel.save(),
-            SelectablePanel::Text(panel) => panel.save(),
-            SelectablePanel::Plot(panel) => panel.save(),
-            SelectablePanel::Image(panel) => panel.save(),
-            SelectablePanel::ImageSegments(panel) => panel.save(),
-            SelectablePanel::Map(panel) => panel.save(),
-            SelectablePanel::Parameter(panel) => panel.save(),
-            SelectablePanel::ManualCalibration(panel) => panel.save(),
-            SelectablePanel::LookAt(panel) => panel.save(),
-            SelectablePanel::Remote(panel) => panel.save(),
-            SelectablePanel::VisionTuner(panel) => panel.save(),
-        };
-        value["_panel_type"] = Value::String(self.to_string());
-
-        value
-    }
-}
-
-impl Widget for &mut SelectablePanel {
-    fn ui(self, ui: &mut Ui) -> eframe::egui::Response {
-        match self {
-            SelectablePanel::BehaviorSimulator(panel) => panel.ui(ui),
-            SelectablePanel::Text(panel) => panel.ui(ui),
-            SelectablePanel::Plot(panel) => panel.ui(ui),
-            SelectablePanel::Image(panel) => panel.ui(ui),
-            SelectablePanel::ImageSegments(panel) => panel.ui(ui),
-            SelectablePanel::Map(panel) => panel.ui(ui),
-            SelectablePanel::Parameter(panel) => panel.ui(ui),
-            SelectablePanel::ManualCalibration(panel) => panel.ui(ui),
-            SelectablePanel::LookAt(panel) => panel.ui(ui),
-            SelectablePanel::Remote(panel) => panel.ui(ui),
-            SelectablePanel::VisionTuner(panel) => panel.ui(ui),
-        }
-    }
-}
-
-impl Display for SelectablePanel {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        let panel_name = match self {
-            SelectablePanel::BehaviorSimulator(_) => BehaviorSimulatorPanel::NAME,
-            SelectablePanel::Text(_) => TextPanel::NAME,
-            SelectablePanel::Plot(_) => PlotPanel::NAME,
-            SelectablePanel::Image(_) => ImagePanel::NAME,
-            SelectablePanel::ImageSegments(_) => ImageSegmentsPanel::NAME,
-            SelectablePanel::Map(_) => MapPanel::NAME,
-            SelectablePanel::Parameter(_) => ParameterPanel::NAME,
-            SelectablePanel::ManualCalibration(_) => ManualCalibrationPanel::NAME,
-            SelectablePanel::LookAt(_) => LookAtPanel::NAME,
-            SelectablePanel::Remote(_) => RemotePanel::NAME,
-            SelectablePanel::VisionTuner(_) => VisionTunerPanel::NAME,
-        };
-        f.write_str(panel_name)
-    }
-}
 
 struct TwixApp {
     nao: Arc<Nao>,
@@ -215,7 +117,7 @@ impl TwixApp {
                     .unwrap()
                     .into()
             }),
-            None => Tree::new(vec![SelectablePanel::Text(TextPanel::new(
+            None => Tree::new(vec![SelectablePanel::TextPanel(TextPanel::new(
                 nao.clone(),
                 None,
             ))
@@ -298,19 +200,7 @@ impl App for TwixApp {
                     }
                     let panel_input = CompletionEdit::new(
                         &mut self.panel_selection,
-                        vec![
-                            "Behavior Simulator".to_string(),
-                            "Text".to_string(),
-                            "Plot".to_string(),
-                            "Image".to_string(),
-                            "Image Segments".to_string(),
-                            "Map".to_string(),
-                            "Parameter".to_string(),
-                            "Manual Calibration".to_string(),
-                            "Look At".to_string(),
-                            "Remote".to_string(),
-                            "Vision Tuner".to_string(),
-                        ],
+                        SelectablePanel::registered(),
                         "Panel",
                     )
                     .ui(ui);
@@ -348,7 +238,7 @@ impl App for TwixApp {
         });
         CentralPanel::default().show(context, |ui| {
             if ui.input_mut(|input| input.consume_key(Modifiers::CTRL, Key::T)) {
-                let tab = SelectablePanel::Text(TextPanel::new(self.nao.clone(), None));
+                let tab = SelectablePanel::TextPanel(TextPanel::new(self.nao.clone(), None));
                 self.tree.push_to_focused_leaf(tab.into());
             }
 
@@ -360,7 +250,7 @@ impl App for TwixApp {
                 .show_add_buttons(true)
                 .show_inside(ui, &mut tab_viewer);
             for node_id in tab_viewer.nodes_to_add_tabs_to {
-                let tab = SelectablePanel::Text(TextPanel::new(self.nao.clone(), None));
+                let tab = SelectablePanel::TextPanel(TextPanel::new(self.nao.clone(), None));
                 let index = self.tree[node_id].tabs_count();
                 self.tree[node_id].insert_tab(index.into(), tab.into());
             }
